Added FIFO for times and scans, added tag-carNum lookup, commonized the database interface, added system testing

+++ CarRegistrationListener/CarRegistrationListener.py
@@ -7,10 +7,12 @@
 #                                                                                     #
 ############################################################ Pittsburgh Shootout LLC ##
 
-from config import Config
 import time
 import atexit
-from database_helper import *
+import os, sys
+sys.path.insert(0, os.path.abspath(".."))
+from Common.config import Config
+from Common.database_helper import *
 import paho.mqtt.client as paho
 import json
 import datetime
@@ -21,44 +23,24 @@ def exit_handler():
 	exit(1)
 
 def environment_check():
-	## Check if table exists
-	## If EXISTS, clear rows if desired
-	## If !EXISTS, create
-
 	delete_if_exists = False
+	table_exists = False
+	function_exists = False
+	trigger_exists = False
 
-	print("CHECKING")
-	if(check_table_exist(table_name)):
-		print("TABLE EXISTS")
-		if(delete_if_exists):
-			print("DELETING ROWS")
-			delete_all_table_rows(table_name)
-		else:
-			print("NOT DELETING ROWS")
-	else:
-		print("CREATING")
-		sql_cmd ="""
-			CREATE TABLE {table_name}(
-				id SERIAL PRIMARY KEY,
-				scan_time VARCHAR(255),
-				tag_number  VARCHAR(255),
-				car_number VARCHAR(255),
-				team_name VARCHAR(255),
-				created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-				updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-				UNIQUE (car_number)
-			)
-			""".format(table_name=table_name)
-		create_table(sql_cmd)
+	table_exists  = check_table_exist(table_name)
+	function_name = table_name + "_set_timestamp"
+	trigger_name  = table_name + "_trigger_set_timestamp"
+	function_exists = check_function_exists(function_name) 
+	trigger_exists  = check_trigger_exists(trigger_name)
+	if(table_exists and delete_if_exists):
+		print("Deleting all rows from " + table_name + "...")
+		delete_all_table_rows(table_name)
+	print(table_name    + " Present: " + str(table_exists))
+	print(function_name + " Present: " + str(function_name))
+	print(trigger_name  + " Present: " + str(trigger_exists))
 
-	## Check if function and trigger are present. If not, create
-	function_name = table_name + "_trigger_set_timestamp"
-	trigger_name = table_name + "_set_timestamp"
-	if(not check_function_exists(function_name) and not check_trigger_exists(trigger_name)):
-		print("Function or trigger do not exist, creating...")
-		create_timestamp_function(function_name)
-		create_timestamp_trigger(table_name, function_name, trigger_name)
-	return True
+	return (table_exists and function_exists and trigger_exists)
 
 def create_mqtt_connection():
 	def on_connect(client, userdata, flags, reason_code, properties):
@@ -70,7 +52,7 @@ def create_mqtt_connection():
 	client = paho.Client(paho.CallbackAPIVersion.VERSION2,client_id=client_id)
 	client.username_pw_set(Config.MQTTUSERNAME, Config.MQTTPASSWORD)
 	client.on_connect = on_connect
-	client.connect(Config.MQTTBROKER, Config.MQTTPORT)
+	client.connect(Config.CarRegistration.MQTTBROKER, Config.MQTTPORT)
 	return client
 
 def sub_handler(client, userdata, msg):
@@ -98,7 +80,7 @@ if __name__ == '__main__':
 		print("Database Schema not correct. Exiting.")
 		atexit.register(exit_handler)
 	
-	client_id = Config.MQTTCLIENTID
+	client_id = Config.CarRegistration.MQTTCLIENTID
 	client = create_mqtt_connection()
 	client.on_message = sub_handler
 	client.loop_start()
